uproszczono model komunikacji
